@@ -21,17 +21,12 @@
         }
     },
     "require": {
-        "php": ">=5.5.0",
-        "guzzlehttp/guzzle": "~5.0|~6.0",
-        "illuminate/support": "~5.0"
+        "php": "^8.0",
+        "guzzlehttp/guzzle": "^7.0",
+        "laravel/framework": "^10.0"
     },
     "require-dev": {
-        "phpspec/phpspec": "~2.2"
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "1.0-dev"
-        }
-    },
-    "minimum-stability": "dev"
+        "nunomaduro/larastan": "^2.5.1",
+        "orchestra/testbench": "^8.0.7"
+    }
 }
