@@ -11,49 +11,31 @@ use Illuminate\Support\Collection;
 use InvalidArgumentException;
 use Psr\Http\Message\ResponseInterface;
 
-/**
- * @method Collection get($resource, array $options = [])
- * @method Collection head($resource, array $options = [])
- * @method Collection put($resource, array $options = [])
- * @method Collection post($resource, array $options = [])
- * @method Collection patch($resource, array $options = [])
- * @method Collection delete($resource, array $options = [])
- */
 class Mailchimp
 {
+    /**
+     * Endpoint for Mailchimp API v3.
+     */
+    private string $endpoint = 'https://us1.api.mailchimp.com/3.0/';
+
+    private string $apikey;
+
+    private Client $client;
 
     /**
-     * Endpoint for Mailchimp API v3
-     *
-     * @var string
+     * @var string[]
      */
-    private $endpoint = 'https://us1.api.mailchimp.com/3.0/';
+    private array $allowedMethods = ['get', 'head', 'put', 'post', 'patch', 'delete'];
 
     /**
-     * @var string
+     * @var array<string,mixed>
      */
-    private $apikey;
+    public array $options = [];
 
     /**
-     * @var Client
+     * @param array<string,mixed> $clientOptions
      */
-    private $client;
-
-    /**
-     * @var array
-     */
-    private $allowedMethods = ['get', 'head', 'put', 'post', 'patch', 'delete'];
-
-    /**
-     * @var array
-     */
-    public $options = [];
-
-    /**
-     * @param string $apikey
-     * @param array $clientOptions
-     */
-    public function __construct($apikey = '', $clientOptions = [])
+    public function __construct(string $apikey = '', array $clientOptions = [])
     {
         $this->apikey = $apikey;
         $this->client = new Client($clientOptions);
@@ -61,20 +43,16 @@ class Mailchimp
         $this->detectEndpoint($this->apikey);
 
         $this->options['headers'] = [
-            'Authorization' => 'apikey ' . $this->apikey
+            'Authorization' => 'apikey '.$this->apikey,
         ];
     }
 
     /**
-     * @param string $resource
-     * @param array $arguments
-     * @param string $method
-     * @return string
      * @throws Exception
      */
-    public function request($resource, $arguments = [], $method = 'GET')
+    public function request(string $resource, array $arguments = [], string $method = 'GET'): Collection
     {
-        if ( ! $this->apikey) {
+        if (! $this->apikey) {
             throw new Exception('Please provide an API key.');
         }
 
@@ -83,44 +61,31 @@ class Mailchimp
 
     /**
      * Enable proxy if needed.
-     *
-     * @param string $host
-     * @param int $port
-     * @param bool $ssl
-     * @param string $username
-     * @param string $password
-     * @return string
      */
     public function setProxy(
-        $host,
-        $port,
-        $ssl = false,
-        $username = null,
-        $password = null
-    ) {
+        string $host,
+        int $port,
+        bool $ssl = false,
+        ?string $username = null,
+        ?string $password = null
+    ): string {
         $scheme = ($ssl ? 'https://' : 'http://');
 
-        if ( ! is_null($username)) {
+        if (! is_null($username)) {
             return $this->options['proxy'] = sprintf('%s%s:%s@%s:%s', $scheme, $username, $password, $host, $port);
         }
 
         return $this->options['proxy'] = sprintf('%s%s:%s', $scheme, $host, $port);
     }
 
-    /**
-     * @return string
-     */
-    public function getEndpoint()
+    public function getEndpoint(): string
     {
         return $this->endpoint;
     }
 
-    /**
-     * @param $apikey
-     */
-    public function detectEndpoint($apikey)
+    public function detectEndpoint(string $apikey): void
     {
-        if ( ! strstr($apikey, '-')) {
+        if (! strstr($apikey, '-')) {
             throw new InvalidArgumentException('There seems to be an issue with your apikey. Please consult Mailchimp');
         }
 
@@ -128,10 +93,7 @@ class Mailchimp
         $this->endpoint = str_replace('us1', $dc, $this->endpoint);
     }
 
-    /**
-     * @param string $apikey
-     */
-    public function setApiKey($apikey)
+    public function setApiKey(string $apikey): void
     {
         $this->detectEndpoint($apikey);
 
@@ -139,17 +101,13 @@ class Mailchimp
     }
 
     /**
-     * @param string $resource
-     * @param array $arguments
-     * @param string $method
-     * @return string
      * @throws Exception
      */
-    private function makeRequest($resource, $arguments, $method)
+    private function makeRequest(string $resource, array $arguments, string $method): Collection
     {
         try {
             $options = $this->getOptions($method, $arguments);
-            $response = $this->client->{$method}($this->endpoint . $resource, $options);
+            $response = $this->client->{$method}($this->endpoint.$resource, $options);
 
             $collection = new Collection(
                 json_decode($response->getBody())
@@ -166,7 +124,7 @@ class Mailchimp
             $response = $e->getResponse();
 
             if ($response instanceof ResponseInterface) {
-                throw new Exception($e->getResponse()->getBody());
+                throw new Exception($response->getBody());
             }
 
             throw new Exception($e->getMessage());
@@ -174,11 +132,9 @@ class Mailchimp
     }
 
     /**
-     * @param string $method
-     * @param array $arguments
-     * @return array
+     * @return array<string,mixed>
      */
-    private function getOptions($method, $arguments)
+    private function getOptions(string $method, array $arguments): array
     {
         unset($this->options['json'], $this->options['query']);
 
@@ -196,23 +152,20 @@ class Mailchimp
     }
 
     /**
-     * @param string $method
-     * @param array $arguments
-     * @return Collection
      * @throws Exception
      */
-    public function __call($method, $arguments)
+    public function __call(string $method, array $arguments): Collection
     {
         if (count($arguments) < 1) {
             throw new InvalidArgumentException('Magic request methods require a URI and optional options array');
         }
 
-        if ( ! in_array($method, $this->allowedMethods)) {
-            throw new BadMethodCallException('Method "' . $method . '" is not supported.');
+        if (! in_array($method, $this->allowedMethods)) {
+            throw new BadMethodCallException('Method "'.$method.'" is not supported.');
         }
 
         $resource = $arguments[0];
-        $options = isset($arguments[1]) ? $arguments[1] : [];
+        $options = $arguments[1] ?? [];
 
         return $this->request($resource, $options, $method);
     }
